refactor ops out of splinter.rs to prepare for more ops
