+++ src/splinter.rs
@@ -1,11 +1,10 @@
-use std::{fmt::Debug, ops::Deref};
+use std::fmt::Debug;
 
 use bytes::Bytes;
 
 use crate::{
-    Cut, Encodable, Merge, Optimizable, SplinterRef,
+    Encodable, Optimizable, SplinterRef,
     codec::{encoder::Encoder, footer::Footer},
-    cow::CowSplinter,
     level::High,
     partition::Partition,
     traits::{PartitionRead, PartitionWrite},
@@ -87,6 +86,21 @@ impl Splinter {
     /// ```
     pub fn encode_to_splinter_ref(&self) -> SplinterRef<Bytes> {
         SplinterRef { data: self.encode_to_bytes() }
+    }
+
+    #[inline(always)]
+    pub(crate) fn new(inner: Partition<High>) -> Self {
+        Self(inner)
+    }
+
+    #[inline(always)]
+    pub(crate) fn inner(&self) -> &Partition<High> {
+        &self.0
+    }
+
+    #[inline(always)]
+    pub(crate) fn inner_mut(&mut self) -> &mut Partition<High> {
+        &mut self.0
     }
 }
 
@@ -320,70 +334,6 @@ impl Optimizable for Splinter {
     #[inline]
     fn optimize(&mut self) {
         self.0.optimize();
-    }
-}
-
-impl Merge for Splinter {
-    fn merge(&mut self, rhs: &Self) {
-        self.0.merge(&rhs.0)
-    }
-}
-
-impl<B: Deref<Target = [u8]>> Merge<SplinterRef<B>> for Splinter {
-    fn merge(&mut self, rhs: &SplinterRef<B>) {
-        self.0.merge(&rhs.load_unchecked())
-    }
-}
-
-impl<B: Deref<Target = [u8]>> Merge<CowSplinter<B>> for Splinter {
-    fn merge(&mut self, rhs: &CowSplinter<B>) {
-        match rhs {
-            CowSplinter::Ref(splinter_ref) => self.merge(splinter_ref),
-            CowSplinter::Owned(splinter) => self.merge(splinter),
-        }
-    }
-}
-
-impl Cut for Splinter {
-    type Out = Self;
-
-    fn cut(&mut self, rhs: &Self) -> Self::Out {
-        Self(self.0.cut(&rhs.0))
-    }
-}
-
-impl<B: Deref<Target = [u8]>> Cut<SplinterRef<B>> for Splinter {
-    type Out = Self;
-
-    fn cut(&mut self, rhs: &SplinterRef<B>) -> Self::Out {
-        Self(self.0.cut(&rhs.load_unchecked()))
-    }
-}
-
-impl<B: Deref<Target = [u8]>> Cut<CowSplinter<B>> for Splinter {
-    type Out = Self;
-
-    fn cut(&mut self, rhs: &CowSplinter<B>) -> Self::Out {
-        match rhs {
-            CowSplinter::Ref(splinter_ref) => self.cut(splinter_ref),
-            CowSplinter::Owned(splinter) => self.cut(splinter),
-        }
-    }
-}
-
-impl<B: Deref<Target = [u8]>> PartialEq<SplinterRef<B>> for Splinter {
-    #[inline]
-    fn eq(&self, other: &SplinterRef<B>) -> bool {
-        self.0 == other.load_unchecked()
-    }
-}
-
-impl<B: Deref<Target = [u8]>> PartialEq<CowSplinter<B>> for Splinter {
-    fn eq(&self, other: &CowSplinter<B>) -> bool {
-        match other {
-            CowSplinter::Ref(splinter_ref) => self.eq(splinter_ref),
-            CowSplinter::Owned(splinter) => self.eq(splinter),
-        }
     }
 }
